[x64] Shader compiler (dx_proxy, remoteshadercompile, shadercompile, vmpi)

+++ tracker/common/msgbuffer.cpp
@@ -350,7 +350,7 @@ char *CMsgBuffer::ReadString (void)
 			break;
 		string[l] = c;
 		l++;
-	} while ( l < sizeof(string)-1 );
+	} while ( l < static_cast<int>(sizeof(string))-1 );
 	
 	string[ l ] = 0;
